Fixed the bug when INSTALL_DOCKER_CLIENT is set the Vagrantfile was ignoring it

+++ Packer/Vagrantfiles/vagrantfile.rb
@@ -1,30 +1,6 @@
 # -*- mode: ruby -*-
 # vi: set ft=ruby :
 
-module LocalCommand
-  class Config < Vagrant.plugin("2", :config)
-      attr_accessor :command
-  end
-
-  class Plugin < Vagrant.plugin("2")
-      name "local_shell"
-
-      config(:local_shell, :provisioner) do
-          Config
-      end
-
-      provisioner(:local_shell) do
-          Provisioner
-      end
-  end
-
-  class Provisioner < Vagrant.plugin("2", :provisioner)
-      def provision
-          result = system "#{config.command}"
-      end
-  end
-end
-
 module OS
   def OS.windows?
       (/cygwin|mswin|mingw|bccwin|wince|emx/ =~ RUBY_PLATFORM) != nil
@@ -108,28 +84,42 @@ Vagrant.configure(2) do |config|
     vb.customize ["modifyvm", :id, "--audio", "none"]
     vb.customize ["modifyvm", :id, "--uart1", "off"] #Disconnect serial port to permit box up on windows/non-unixlike devices
     vb.customize ['modifyvm', :id, '--vrde', 'off']
-    override.vm.provision "shell", inline: "/usr/bin/deploy-docker-files.sh #{Host_OS} virtualbox"
+
+    if ENV['INSTALL_DOCKER_CLIENT'].to_s.downcase != "no"
+      override.vm.provision "shell", inline: "/usr/bin/deploy-docker-files.sh #{Host_OS} virtualbox"
+    end
+
   end
 
-  config.vm.provider "hyperv" do |vb, override|
-    override.vm.provision "shell", inline: "/usr/bin/deploy-docker-files.sh #{Host_OS} hyperv"
+  config.vm.provider "hyperv" do |hv, override|
+    hv.memory = 512
+    hv.cpus = 2
+    if ENV['INSTALL_DOCKER_CLIENT'].to_s.downcase != "no"
+      override.vm.provision "shell", inline: "/usr/bin/deploy-docker-files.sh #{Host_OS} hyperv"
+    end
   end
 
-  config.vm.provider "parallels" do |vb, override|
-    override.vm.provision "shell", inline: "/usr/bin/deploy-docker-files.sh #{Host_OS} parallels"
+  config.vm.provider "parallels" do |pl, override|
+    pl.memory = 512
+    pl.cpus = 2
+    if ENV['INSTALL_DOCKER_CLIENT'].to_s.downcase != "no"
+      override.vm.provision "shell", inline: "/usr/bin/deploy-docker-files.sh #{Host_OS} parallels"
+    end
   end
 
-
-  config.trigger.before :destroy do |trigger|
-    trigger.name = "Uninstalling docker client"
-    trigger.run = {inline: $uninstall_docker_client}
-    trigger.run_remote = {inline: "rm -fr /vagrant/installer"}
+  if ENV['INSTALL_DOCKER_CLIENT'].to_s.downcase != "no"
+    config.trigger.before :destroy do |trigger|
+      trigger.name = "Uninstalling docker client"
+      trigger.run = {inline: $uninstall_docker_client}
+      trigger.run_remote = {inline: "rm -fr /vagrant/installer"}
+    end
   end
 
-  config.trigger.after :up do |trigger|
-    trigger.name = "Installing docker client"
-    trigger.run = {inline: $install_docker_client}
+  if ENV['INSTALL_DOCKER_CLIENT'].to_s.downcase != "no"
+    config.trigger.after :up do |trigger|
+      trigger.name = "Installing docker client"
+      trigger.run = {inline: $install_docker_client}
+    end
   end
 
-  #config.vm.provision "install-docker-cli", type: "local_shell", command: $install_docker_client
 end
